Made the apn production setting more visable
It took me hours to figure out why I was getting BadDeviceToken in production while on my test iPhone everything was working great. I had to go to the node-apn package to figure this out. 

The production setting should be shown in the settings (as it is inside the node-apn readme) so people can see they have to toggle it.